+++ src/main/java/com/kynsoft/report/applications/command/produccionterminada/delete/DeleteProduccionTerminadaMessage.java
@@ -10,10 +10,10 @@ public class DeleteProduccionTerminadaMessage implements ICommandMessage {
     private final UUID id;
     private final Double stockAnterior;
     private final Double stockNuevo;
-    private final Integer cantidadRevertida;
+    private final Double cantidadRevertida;
     private final String command = "DELETE_PRODUCCION_TERMINADA";
 
-    public DeleteProduccionTerminadaMessage(UUID id, Double stockAnterior, Double stockNuevo, Integer cantidadRevertida) {
+    public DeleteProduccionTerminadaMessage(UUID id, Double stockAnterior, Double stockNuevo, Double cantidadRevertida) {
         this.id = id;
         this.stockAnterior = stockAnterior;
         this.stockNuevo = stockNuevo;
